Move tunnel-related code to root
Subfolders are safe to mess with. Root objects, not quite.
